Add all switches to sensors, manually set speed at start
Assumed that all cases for the switches are thought out, although no semaphores are implented, therefore there might be some cases where switches won't turn automatically.
By reaching a station, the speed parameter gets inverted and if a person;
sets  one or both of the speed parameters to 0,starts the program, assign it a new speed manually, the program  will cause the speed to become 0 when reaching a station.